Refactor test files: rename and reorganize NoSQL database test utilities
