+++ src/core/tests/no-sql-db-triggers.utils.test.ts
@@ -4,7 +4,7 @@ import {
   dbTestRootPath,
   pathTestDocument,
   TestDocumentData,
-} from "./no-sql-db-spec";
+} from "./no-sql-db.utils.test";
 
 export function testNoSqlDbTriggers(db: INoSqlDatabase) {
   describe("NoSqlDatabase - Triggers", () => {
