+++ src/provider/fake/database.fake.test.ts
@@ -1,4 +1,4 @@
-import { testNoSqlDb } from "../../core/tests/no-sql-db-spec";
+import { testNoSqlDb } from "../../core/tests/no-sql-db.utils.test";
 import { createNoSqlDatabaseTesting } from "./database.fake";
 
 testNoSqlDb(createNoSqlDatabaseTesting());
